refactor utils to not depend on director imports
