+++ modules/spartan/perception/cameratransform.py
@@ -23,6 +23,8 @@ import bot_core as lcmbotcore
 
 # spartan
 import spartan.utils.utils as spartanUtils
+import spartan.utils.director_utils as spartanDirectorUtils
+
