+++ modules/spartan/utils/ros_utils.py
@@ -1,17 +1,34 @@
+# system
 import yaml
 import time
 import random
 import os
 import math
 
-from geometry_msgs.msg import PoseStamped
-from geometry_msgs.msg import Transform
-
-from director import transformUtils
-
 
 # ROS
 import rospy
+import geometry_msgs.msg
+
+# spartan
+import spartan.utils.utils as spartanUtils
+
+
+
+def ROSTransformMsgFromPose(d):
+    msg = geometry_msgs.msg.Transform()
+    msg.translation.x = d['translation']['x']
+    msg.translation.y = d['translation']['y']
+    msg.translation.z = d['translation']['z']
+
+    quatDict = spartanUtils.getQuaternionFromDict(d)
+
+    msg.rotation.w = quatDict['w']
+    msg.rotation.x = quatDict['x']
+    msg.rotation.y = quatDict['y']
+    msg.rotation.z = quatDict['z']
+
+    return msg
 
 class SimpleSubscriber(object):
     def __init__(self, topic, messageType, externalCallback=None):
